Created CSF directory
