+++ packages/lib/src/components/internal/SecuredFields/lib/CSF/CSF.ts
@@ -1,24 +1,25 @@
 import { handleConfig } from '../configuration/handleConfig';
 import { configureCallbacks } from '../configuration/configureCallbacks';
-import { handleProcessBrand } from './utils/processBrand';
-import { handleValidation } from './handleValidation';
-import { handleEncryption } from './handleEncryption';
-import { createSecuredFields, createNonCardSecuredFields, createCardSecuredFields, setupSecuredField } from './createSecuredFields';
-import { setFocusOnFrame } from './utils/iframes/setFocusOnFrame';
-import { postMessageToAllIframes } from './utils/iframes/postMessageToAllIframes';
-import { destroySecuredFields } from './destroySecuredFields';
-import { processAutoComplete } from './utils/processAutoComplete';
-import { handleFocus } from './utils/iframes/handleFocus';
-import { handleIframeConfigFeedback } from './utils/iframes/handleIframeConfigFeedback';
-import { isConfigured } from './utils/isConfigured';
-import { assessFormValidity } from './utils/validateForm';
-import { handleBinValue } from './utils/handleBinValue';
-import { handleBrandFromBinLookup, sendBrandToCardSF, sendExpiryDatePolicyToSF } from './utils/handleBrandFromBinLookup';
-import handleAdditionalFields from './utils/registerAdditionalField';
-import tabHandlers from './utils/tabbing/handleTab';
-import postMessageToIframe from './utils/iframes/postMessageToIframe';
+import { handleProcessBrand } from '../core/utils/processBrand';
+import { handleValidation } from '../core/handleValidation';
+import { handleEncryption } from '../core/handleEncryption';
+import { createSecuredFields, createNonCardSecuredFields, createCardSecuredFields, setupSecuredField } from '../core/createSecuredFields';
+import { setFocusOnFrame } from '../core/utils/iframes/setFocusOnFrame';
+import { postMessageToAllIframes } from '../core/utils/iframes/postMessageToAllIframes';
+import { destroySecuredFields } from '../core/destroySecuredFields';
+import { processAutoComplete } from '../core/utils/processAutoComplete';
+import { handleFocus } from '../core/utils/iframes/handleFocus';
+import { handleIframeConfigFeedback } from '../core/utils/iframes/handleIframeConfigFeedback';
+import { isConfigured } from '../core/utils/isConfigured';
+import { assessFormValidity } from '../core/utils/validateForm';
+import { handleBinValue } from '../core/utils/handleBinValue';
+import { handleBrandFromBinLookup, sendBrandToCardSF, sendExpiryDatePolicyToSF } from '../core/utils/handleBrandFromBinLookup';
+import handleAdditionalFields from '../core/utils/registerAdditionalField';
+import tabHandlers from '../core/utils/tabbing/handleTab';
+import postMessageToIframe from '../core/utils/iframes/postMessageToIframe';
 import AbstractCSF from './AbstractCSF';
-import { CSFReturnObject, CSFSetupObject, StylesObject, CbObjOnAdditionalSF, CSFStateObject } from '../types';
+import { CSFReturnObject, CSFSetupObject, CSFStateObject } from './types';
+import { StylesObject, CbObjOnAdditionalSF } from '../types';
 import * as logger from '../utilities/logger';
 import { selectOne } from '../utilities/dom';
 import { BinLookupResponse } from '../../../../Card/types';
