+++ packages/lib/src/components/internal/SecuredFields/lib/CSF/index.ts
@@ -1,0 +1,3 @@
+import csf from './initCSF';
+
+export default csf;
