--- packages/lib/src/components/internal/SecuredFields/lib/index.ts
@@ -1,3 +1,0 @@
-import csf from './core/initCSF';
-
-export default csf;
